# sql,tree,eval: update oid functions to use uint32