--- a/pkg/sql/opt/partition/BUILD.bazel
+++ b/pkg/sql/opt/partition/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//pkg/sql/types",
         "//pkg/util",
         "@com_github_cockroachdb_errors//:errors",
+        "@com_github_lib_pq//oid",
     ],
 )