fixes https://github.com/cockroachdb/cockroach/issues/41904

See individual commits.