@yuzefovich 2022-10-04
`c48196c300087cc1` is responsible for a regression on a micro-benchmark:
```
BENCHTIMEOUT=24h PKG=./pkg/bench BENCHES=SQL/MultinodeCockroach/Upsert/count=1000 ./scripts/bench c48196c300087cc1^ c48196c300087cc1
...
+ ./dev bench ./pkg/bench --timeout=24h --filter=SQL/MultinodeCockroach/Upsert/count=1000 --count=10 --bench-mem -v --stream-output --ignore-cache
Switching to c48196c300087cc1
+ ./dev bench ./pkg/bench --timeout=24h --filter=SQL/MultinodeCockroach/Upsert/count=1000 --count=10 --bench-mem -v --stream-output --ignore-cache
name                                         old time/op    new time/op    delta
SQL/MultinodeCockroach/Upsert/count=1000-24    23.8ms ± 9%    25.9ms ± 9%   +8.78%  (p=0.003 n=10+10)

name                                         old alloc/op   new alloc/op   delta
SQL/MultinodeCockroach/Upsert/count=1000-24    11.2MB ±29%    12.2MB ±12%     ~     (p=0.089 n=10+10)

name                                         old allocs/op  new allocs/op  delta
SQL/MultinodeCockroach/Upsert/count=1000-24     54.9k ±14%     79.3k ±10%  +44.47%  (p=0.000 n=10+10)
```
Is this concerning?